session: Rename files to follow the prefix-name rule
Renames the session files to xdp-session.
