+++ src/clipboard.c
@@ -24,7 +24,7 @@
 
 #include "clipboard.h"
 #include "remote-desktop.h"
-#include "session.h"
+#include "xdp-session.h"
 #include "xdp-dbus.h"
 #include "xdp-impl-dbus.h"
 #include "xdp-utils.h"
