+++ src/global-shortcuts.c
@@ -27,7 +27,7 @@
 
 #include "global-shortcuts.h"
 #include "xdp-request.h"
-#include "session.h"
+#include "xdp-session.h"
 #include "permissions.h"
 #include "xdp-dbus.h"
 #include "xdp-impl-dbus.h"
