+++ src/meson.build
@@ -94,7 +94,6 @@ xdg_desktop_portal_sources = files(
   'screen-cast.c',
   'screenshot.c',
   'secret.c',
-  'session.c',
   'settings.c',
   'trash.c',
   'wallpaper.c',
@@ -104,6 +103,7 @@ xdg_desktop_portal_sources = files(
   'xdp-call.c',
   'xdp-portal-impl.c',
   'xdp-request.c',
+  'xdp-session.c',
   'xdp-session-persistence.c',
 )
