+++ src/xdp-session.c
@@ -18,7 +18,7 @@
  *
  */
 
-#include "session.h"
+#include "xdp-session.h"
 #include "xdp-request.h"
 #include "xdp-call.h"
