+++ tests/backend/xdp-session.c
@@ -19,7 +19,7 @@
  *
  */
 
-#include "session.h"
+#include "xdp-session.h"
 
 enum
 {
