+++ tests/meson.build
@@ -63,7 +63,6 @@ test(
 test_backends = executable(
   'test-backends',
   'backend/test-backends.c',
-  'backend/session.c',
   'backend/access.c',
   'backend/account.c',
   'backend/appchooser.c',
@@ -78,6 +77,7 @@ test_backends = executable(
   'backend/settings.c',
   'backend/wallpaper.c',
   'backend/xdp-request.c',
+  'backend/xdp-session.c',
   'glib-backports.c',
   document_portal_built_sources,
   document_portal_built_sources,
